Fix unsafe uses of unsafe.Pointer (#1438)
Per rule 4 of unsafe.Pointer usage, conversion of a unsafe.Pointer
to a uinptr to pass to the syscall.Syscall family should only be done in
the argument list. We had a couple spots where we were passing it in as
an argument to a small wrapper function until it reached the underlying
syscall.Syscall*.

https://pkg.go.dev/unsafe#Pointer

+++ internal/jobcontainers/jobcontainer.go
@@ -731,7 +731,7 @@ func systemProcessInformation() ([]*winapi.SYSTEM_PROCESS_INFORMATION, error) {
 		systemProcInfo = (*winapi.SYSTEM_PROCESS_INFORMATION)(unsafe.Pointer(&b[0]))
 		status := winapi.NtQuerySystemInformation(
 			winapi.SystemProcessInformation,
-			uintptr(unsafe.Pointer(systemProcInfo)),
+			unsafe.Pointer(systemProcInfo),
 			size,
 			&size,
 		)
